Create email_spam_using_ML
Email Spam Classification is a machine learning project that involves training a model to automatically distinguish between spam (unwanted or junk email) and ham (legitimate email) using a labeled dataset. The goal is to build a system that can accurately classify incoming emails and help reduce spam in a user’s inbox.